refactor: AppointmentController에서 AppointmentService 서비스 레이어 추출
- AppointmentService 신규 생성: getByDateRange, getById, create, updateStatus(suspend), cancel(suspend)
- AppointmentController 의존성 5개 → 2개(AppointmentService, ClinicTimezoneService)로 축소
- 비즈니스 로직(상태머신, 이벤트 발행, 이력 저장) Controller → Service로 이동
- parseEvent() 내부 함수 AppointmentService.kt로 이동
